+++ NEWS.md
@@ -8,6 +8,8 @@
 
 - Remove tests relying on non-convergence of the FOMC fit to the FOCUS A dataset as this is platform dependent (revealed by the new additional tests on CRAN, thanks!)
 
+- Increase test tolerance for some parameter comparisons that also proved to be platform dependent
+
 # mkin 1.0.0
 
 ## General